re-arrange some controls on small screens
