+++ src/navigation/Components/LightToggleProfileMenu.js
@@ -14,6 +14,7 @@ import { networkStatusSelector } from "../../redux/Selectors";
 import { logoutUser } from "../../redux/login/LoginActions";
 import SyncStatusCircleLoader from "./SyncStatusCircleLoader";
 import UserFeedbackDialog from "./UserFeedbackDialog";
+import { useTheme, useMediaQuery } from "@mui/material";
 
 function LightToggleProfileMenu() {
     const whoami = useSelector(getWhoami);
@@ -21,6 +22,8 @@ function LightToggleProfileMenu() {
     const [feedbackOpen, setFeedbackOpen] = React.useState(false);
     const dispatch = useDispatch();
     const networkStatus = useSelector(networkStatusSelector);
+    const theme = useTheme();
+    const isSm = useMediaQuery(theme.breakpoints.down("sm"));
 
     return (
         <Stack
@@ -29,17 +32,18 @@ function LightToggleProfileMenu() {
             alignItems={"center"}
             spacing={1}
         >
-            <Box sx={{ width: 40 }}>
+            <Box>
                 <SyncStatusCircleLoader />
-                {process.env.REACT_APP_DEMO_MODE !== "true" && !networkStatus && (
-                    <Tooltip title={"You are working offline"}>
-                        <IconButton size="large">
-                            <SignalWifiOff />
-                        </IconButton>
-                    </Tooltip>
-                )}
+                {process.env.REACT_APP_DEMO_MODE !== "true" &&
+                    !networkStatus && (
+                        <Tooltip title={"You are working offline"}>
+                            <IconButton size="large">
+                                <SignalWifiOff />
+                            </IconButton>
+                        </Tooltip>
+                    )}
             </Box>
-            {process.env.REACT_APP_DEMO_MODE !== "true" && (
+            {process.env.REACT_APP_DEMO_MODE !== "true" && !isSm && (
                 <Tooltip title={"Send feedback"}>
                     <IconButton
                         onClick={() => {
@@ -52,7 +56,7 @@ function LightToggleProfileMenu() {
                     </IconButton>
                 </Tooltip>
             )}
-            <div>
+            <React.Fragment>
                 <UserAvatar
                     onClick={(event) => {
                         setAnchorElProfileMenu(event.currentTarget);
@@ -91,7 +95,7 @@ function LightToggleProfileMenu() {
                         Logout
                     </MenuItem>
                 </Menu>
-            </div>
+            </React.Fragment>
             <UserFeedbackDialog
                 onClose={() => setFeedbackOpen(false)}
                 open={feedbackOpen}
