+++ src/scenes/Dashboard/components/DashboardDetailTabs.tsx
@@ -1,38 +1,25 @@
 import React from "react";
-import * as selectionModeActions from "../../../redux/selectionMode/selectionModeActions";
+import RoleViewSelect from "./RoleViewSelect";
 import AddIcon from "@mui/icons-material/Add";
 import Box from "@mui/material/Box";
 import { useDispatch, useSelector } from "react-redux";
-import MenuItem from "@mui/material/MenuItem";
-import { saveDashboardRoleMode } from "../../../utilities";
 import {
     setDashboardFilteredUser,
-    setDashboardTabIndex,
     setGuidedSetupOpen,
-    setRoleView,
 } from "../../../redux/Actions";
 import TaskFilterTextField from "../../../components/TaskFilterTextfield";
-import {
-    Badge,
-    Chip,
-    Fab,
-    FormControl,
-    Hidden,
-    Select,
-    Stack,
-} from "@mui/material";
+import { Fab, Hidden, Stack } from "@mui/material";
 import { useTheme, useMediaQuery } from "@mui/material";
 import {
     dashboardFilteredUserSelector,
     dashboardFilterTermSelector,
     dashboardTabIndexSelector,
     getRoleView,
-    getWhoami,
     guidedSetupOpenSelector,
 } from "../../../redux/Selectors";
 import { clearDashboardFilter } from "../../../redux/dashboardFilter/DashboardFilterActions";
 import * as models from "../../../models";
-import useTaskObserveQueryByStatus from "../../../hooks/useTaskObserveQueryByStatus";
+import DashboardTabs from "./DashboardTabs";
 
 type DashboardDetailTabsProps = {
     disableAddButton?: boolean;
@@ -42,63 +29,14 @@ const DashboardDetailTabs: React.FC<DashboardDetailTabsProps> = ({
     disableAddButton,
 }) => {
     const dispatch = useDispatch();
-    const whoami = useSelector(getWhoami);
     const dashboardFilter = useSelector(dashboardFilterTermSelector);
     const roleView = useSelector(getRoleView);
     const dashboardFilteredUser = useSelector(dashboardFilteredUserSelector);
     const guidedSetupOpen = useSelector(guidedSetupOpenSelector);
     const theme = useTheme();
     const isSm = useMediaQuery(theme.breakpoints.down("sm"));
-    const isMd = useMediaQuery(theme.breakpoints.down("md"));
     const dashboardTabIndex = useSelector(dashboardTabIndexSelector);
 
-    const pendingTasks = useTaskObserveQueryByStatus(models.TaskStatus.PENDING);
-    const pendingLength = pendingTasks?.state?.length || 0;
-
-    const handleChange = (newValue: number) => {
-        dispatch(setDashboardTabIndex(newValue));
-    };
-
-    const tabs = (
-        <Stack spacing={isSm ? 0.5 : 1} direction="row">
-            <Chip
-                key="dashboard-tab-0"
-                data-testid="dashboard-tab-inprogress"
-                aria-label="Dashboard in Progress"
-                label="IN PROGRESS"
-                color={dashboardTabIndex === 0 ? "primary" : "default"}
-                onClick={() => handleChange(0)}
-            />
-            <Chip
-                key="dashboard-tab-1"
-                data-testid="dashboard-tab-completed"
-                aria-label="Dashboard Completed"
-                onClick={() => handleChange(1)}
-                color={dashboardTabIndex === 1 ? "primary" : "default"}
-                label="COMPLETED"
-            />
-            {roleView !== "RIDER" && (
-                <Badge
-                    invisible={pendingLength === 0}
-                    badgeContent={pendingLength}
-                    color="error"
-                >
-                    <Chip
-                        key="dashboard-tab-2"
-                        data-testid="dashboard-tab-pending"
-                        aria-label="Dashboard Pending"
-                        onClick={() => handleChange(2)}
-                        color={dashboardTabIndex === 2 ? "primary" : "default"}
-                        label="PENDING"
-                    />
-                </Badge>
-            )}
-        </Stack>
-    );
-    const clearAllSelectedItems = () => {
-        dispatch(selectionModeActions.clearItems(0));
-        dispatch(selectionModeActions.clearItems(1));
-    };
     const addClearButton =
         !dashboardFilter && !dashboardFilteredUser ? (
             <Fab
@@ -141,7 +79,9 @@ const DashboardDetailTabs: React.FC<DashboardDetailTabsProps> = ({
             justifyContent={"space-between"}
             alignItems={"center"}
         >
-            <Box>{tabs}</Box>
+            <Box>
+                <DashboardTabs />
+            </Box>
             <Hidden mdDown>
                 <TaskFilterTextField sx={{ width: "40%" }} />
             </Hidden>
@@ -155,85 +95,7 @@ const DashboardDetailTabs: React.FC<DashboardDetailTabsProps> = ({
                     {["ALL", models.Role.COORDINATOR].includes(roleView) &&
                         addClearButton}
                 </Hidden>
-                {dashboardTabIndex !== 2 && (
-                    <FormControl variant="outlined">
-                        <Select
-                            sx={{
-                                right: 5,
-                                borderRadius: 2,
-                                "&.Mui-focused .MuiOutlinedInput-notchedOutline":
-                                    {
-                                        borderColor: "orange",
-                                    },
-                                "&:hover .MuiOutlinedInput-notchedOutline": {
-                                    borderColor: "orange",
-                                },
-                            }}
-                            disabled={guidedSetupOpen}
-                            size="small"
-                            data-testId="role-menu"
-                            value={roleView}
-                        >
-                            {whoami.roles.includes(models.Role.COORDINATOR) && (
-                                <MenuItem
-                                    value="ALL"
-                                    onClick={() => {
-                                        if (roleView !== "ALL") {
-                                            dispatch(setRoleView("ALL"));
-                                            saveDashboardRoleMode("ALL");
-                                            clearAllSelectedItems();
-                                        }
-                                    }}
-                                >
-                                    ALL
-                                </MenuItem>
-                            )}
-                            {whoami.roles.includes(models.Role.COORDINATOR) && (
-                                <MenuItem
-                                    value={models.Role.COORDINATOR}
-                                    onClick={() => {
-                                        if (
-                                            roleView !== models.Role.COORDINATOR
-                                        ) {
-                                            dispatch(
-                                                setRoleView(
-                                                    models.Role.COORDINATOR
-                                                )
-                                            );
-                                            saveDashboardRoleMode(
-                                                models.Role.COORDINATOR
-                                            );
-                                            clearAllSelectedItems();
-                                        }
-                                    }}
-                                >
-                                    {isMd ? "COORD" : "COORDINATOR"}
-                                </MenuItem>
-                            )}
-                            {whoami.roles.includes(models.Role.RIDER) && (
-                                <MenuItem
-                                    value={models.Role.RIDER}
-                                    onClick={() => {
-                                        if (roleView !== models.Role.RIDER) {
-                                            dispatch(
-                                                setRoleView(models.Role.RIDER)
-                                            );
-                                            dispatch(
-                                                setDashboardFilteredUser(null)
-                                            );
-                                            saveDashboardRoleMode(
-                                                models.Role.RIDER
-                                            );
-                                            clearAllSelectedItems();
-                                        }
-                                    }}
-                                >
-                                    RIDER
-                                </MenuItem>
-                            )}
-                        </Select>
-                    </FormControl>
-                )}
+                {!isSm && dashboardTabIndex !== 2 && <RoleViewSelect />}
             </Stack>
         </Stack>
     );
